fix sca
